chore: create config.helpers
